Refactor error types
I'm actually not 100% sure if this is the idomatic way to do things.

Basically, we have a few generic errors:
 - AuthenticationError (returned from parse_auth_response)
 - ParseRequestError (returned from parse_response)

Then we have specific Auth errors for each HTTP impl:
 - ClosureAuthError (wraps AuthenticationError, adds a few other variants)
 - ReqwestAuthError (as above)

We also have generic errors for each HTTP impl:
 - ClosureError (returned from the request.send() method)
 - ReqwestError (as above)

Then, finally, we have a global Error type that implemets variants (and From) for all of the variants on the errors above.

This is a bit hairy because each HTTP impl has its own authenticate() function, with different error possibilities. Further, the error type for closure-based HTTP impls can't be known in this library, because the crate user could be using any HTTP library.

At some point in the future, I want to explore using an associated type on "RequestImpl" for this instead. Really, the only difference between ClosureError and ReqwestError is the named type (e.g. ReqwestError = reqwest::Error, and ClosureError = Box<dyn Error>). We might also be able to refactor the Closure impls to accept a struct that implements a trait from this crate, and could let crate users define an associated type to indicate the errors they may return.